cluster/matrix: build provision_matrix_users.py as an OCI image, relocate to provisioners/
Matches the grocy_user_perms/inventree_token_provisioner pattern for
consistency: the Job now pulls a Bazel-built image instead of mounting
the script from a configMapGenerator. This unblocks the source-code
relocation the earlier commit on this branch couldn't do — kustomize's
default load-restrictor (which Flux's kustomize-controller also
enforces) rejects a configMapGenerator file source outside the
kustomization's own root, so as long as the script was ConfigMap-mounted
it had to stay physically inside cluster/k8s/matrix/secrets/. Baking it
into an image removes that coupling entirely.

- cluster/provisioners/matrix_user_provisioner/: provision_matrix_users.py
  + test_provision_matrix_users.py (moved, import path updated) + new
  BUILD.bazel (py_library, py_test, aspect_py_binary, py_image_layer,
  oci_image, oci_load — mirrors grocy_user_perms/BUILD.bazel).
- cluster/k8s/matrix/secrets/: dropped the local BUILD.bazel (no longer
  a Bazel sub-package) and its manifests filegroup workaround in
  cluster/validation/BUILD.bazel; provision-matrix-users-job.yaml now
  pulls ghcr.io/agentydragon/matrix-user-provisioner with an
  $imagepolicy marker instead of mounting a ConfigMap; kustomization.yaml
  drops the configMapGenerator.
- .github/workflows/push-images.yml: new matrix row, test-gated on the
  moved test suite.
- cluster/k8s/flux-image-automation-ghcr/: new ImageRepository +
  ImagePolicy, registered in kustomization.yaml and the GitHub webhook
  receiver (immediate pickup on push instead of the 5m poll).
- cluster/README.md: doc-tree parenthetical now names all three
  provisioners.

Verified with pip-installed pytest + pytest_bazel (bbr/bazel unable to
reach BuildBuddy in this session), buildifier, and `kustomize build` on
the updated cluster/k8s/matrix/secrets (confirms the configMapGenerator
removal doesn't break the kustomization). pre-commit passed except
markdownlint-cli2 (missing from this session's devtools install, a
pre-existing gap; the one touched Markdown line is inside a fenced code
block — SKIP=markdownlint-cli2 used to commit).

BAZEL_TEST_INVOCATIONS=local:11dc7051-546b-4230-a2e2-872cb9a7df0f
