+++ cluster/k8s/flux-image-automation-ghcr/matrix-user-provisioner-image-policy.yaml
@@ -1,0 +1,14 @@
+apiVersion: image.toolkit.fluxcd.io/v1beta2
+kind: ImagePolicy
+metadata:
+  name: matrix-user-provisioner
+  namespace: flux-system
+spec:
+  imageRepositoryRef:
+    name: matrix-user-provisioner
+  filterTags:
+    # Tags pushed by CI: {branch}-YYYYMMDDHHMMSS-{sha7} — alphabetical order == chronological
+    pattern: '^devel-\d{14}-[0-9a-f]{7}$'
+  policy:
+    alphabetical:
+      order: asc
