--- cluster/k8s/matrix/secrets/BUILD.bazel
@@ -1,27 +1,0 @@
-load("//devinfra/python:defs.bzl", "py_library", "py_test")
-
-py_library(
-    name = "provision_matrix_users",
-    srcs = ["provision_matrix_users.py"],
-)
-
-py_test(
-    name = "test_provision_matrix_users",
-    srcs = ["test_provision_matrix_users.py"],
-    deps = [
-        ":provision_matrix_users",
-        "@pypi//pytest",
-        "@pypi//pytest_bazel",
-    ],
-)
-
-# Cluster manifests for cluster/validation's flux-build test (the
-# //cluster/k8s:cluster_all_files glob stops at this package boundary).
-filegroup(
-    name = "manifests",
-    srcs = glob(
-        ["**"],
-        exclude = ["BUILD.bazel"],
-    ),
-    visibility = ["//cluster/validation:__pkg__"],
-)
