+++ cluster/k8s/matrix/secrets/kustomization.yaml
@@ -8,8 +8,3 @@ resources:
   - synapse-admin-credentials.sops.yaml
   - openclaw-matrix-bot-password.sops.yaml
   - provision-matrix-users-job.yaml
-configMapGenerator:
-  - name: provision-matrix-users-script
-    namespace: matrix
-    files:
-      - provision_matrix_users.py
