+++ cluster/k8s/matrix/secrets/provision-matrix-users-job.yaml
@@ -14,8 +14,9 @@ spec:
       restartPolicy: OnFailure
       containers:
         - name: provision
-          image: python:3.13-alpine
-          command: ["python3", "/scripts/provision_matrix_users.py"]
+          # Script baked in via Bazel (//cluster/provisioners/matrix_user_provisioner:image).
+          image: ghcr.io/agentydragon/matrix-user-provisioner:latest # {"$imagepolicy": "flux-system:matrix-user-provisioner"}
+          imagePullPolicy: Always
           env:
             - name: REGISTRATION_SECRET
               valueFrom:
@@ -32,11 +33,3 @@ spec:
                 secretKeyRef:
                   name: openclaw-matrix-bot-password
                   key: password
-          volumeMounts:
-            - name: script
-              mountPath: /scripts
-              readOnly: true
-      volumes:
-        - name: script
-          configMap:
-            name: provision-matrix-users-script
