+++ cluster/provisioners/matrix_user_provisioner/BUILD.bazel
@@ -1,0 +1,52 @@
+"""Matrix/Synapse user provisioner OCI image — registers the admin user and
+upserts the OpenClaw bot via the Synapse admin API.
+"""
+
+load("@aspect_rules_py//py:defs.bzl", "py_image_layer", aspect_py_binary = "py_binary")
+load("@rules_oci//oci:defs.bzl", "oci_image", "oci_load")
+load("//devinfra/python:defs.bzl", "py_library", "py_test")
+load("//props:oci.bzl", "py_image_entrypoint", "py_image_env")
+
+py_library(
+    name = "provision_matrix_users",
+    srcs = ["provision_matrix_users.py"],
+)
+
+py_test(
+    name = "test_provision_matrix_users",
+    srcs = ["test_provision_matrix_users.py"],
+    deps = [
+        ":provision_matrix_users",
+        "@pypi//pytest",
+        "@pypi//pytest_bazel",
+    ],
+)
+
+aspect_py_binary(
+    name = "provision_matrix_users_bin",
+    main = "provision_matrix_users.py",
+    deps = [":provision_matrix_users"],
+)
+
+py_image_layer(
+    name = "layers",
+    binary = ":provision_matrix_users_bin",
+    group = "1000",
+    owner = "1000",
+)
+
+oci_image(
+    name = "image",
+    base = "@debian_slim_linux_amd64",
+    entrypoint = py_image_entrypoint("provision_matrix_users_bin"),
+    env = py_image_env(binary_name = "provision_matrix_users_bin"),
+    labels = {"org.opencontainers.image.source": "https://github.com/agentydragon/ducktape"},
+    tars = [":layers"],
+    user = "1000:1000",
+)
+
+oci_load(
+    name = "load",
+    image = ":image",
+    repo_tags = ["matrix-user-provisioner:latest"],
+)
